Refactored ClimateController to Vertical Slice Architecture
