+++ WeatherForecast/VerticalSlices/ClimateRequest.cs
@@ -1,0 +1,4 @@
+﻿namespace WeatherForecast.VerticalSlices
+{
+    public record ClimateRequest(string Location, int LowTemperature, int HighTemperature);
+}
